WIP: SC done, Interface lacking DB
Remember to make the insert in a way that allows to log the tx hash and
blocktime, as well as access control to admin. Maybe implement a webhook
such that when user transfers ownership, supabase relays manager.
How possible is it to make user=public_key and pwd=private_key?